Split up views.py
