+++ members/views/PersonUpdate.py
@@ -1,0 +1,28 @@
+import uuid
+
+from django.core.urlresolvers import reverse
+from django.http import HttpResponseRedirect
+from django.shortcuts import render, get_object_or_404
+
+from members.forms import PersonForm
+from members.models.person import Person
+
+from members.views.UpdatePersonFromForm import UpdatePersonFromForm
+
+def PersonUpdate(request, unique, id):
+    try:
+        unique = uuid.UUID(unique)
+    except ValueError:
+        return HttpResponseBadRequest("Familie id er ugyldigt")
+
+    person = get_object_or_404(Person, pk=id)
+    if person.family.unique != unique:
+        raise Http404("Person eksisterer ikke")
+    if request.method == 'POST':
+        form = PersonForm(request.POST, instance=person)
+        if form.is_valid():
+            UpdatePersonFromForm(person,form)
+            return HttpResponseRedirect(reverse('family_detail', args=[person.family.unique]))
+    else:
+        form = PersonForm(instance=person)
+    return render(request, 'members/person_create_or_update.html', {'form': form, 'person': person})
