+++ members/views/__init__.py
@@ -1,0 +1,16 @@
+from members.views.FamilyDetails import FamilyDetails
+from members.views.ConfirmFamily import ConfirmFamily
+from members.views.PersonCreate import PersonCreate
+from members.views.PersonUpdate import PersonUpdate
+from members.views.WaitingListSetSubscription import WaitingListSetSubscription
+from members.views.DeclineInvitation import DeclineInvitation
+from members.views.ActivitySignup import ActivitySignup
+from members.views.UpdatePersonFromForm import UpdatePersonFromForm
+from members.views.EntryPage import EntryPage
+from members.views.volunteerSignup import volunteerSignup
+from members.views.loginEmailSent import loginEmailSent
+from members.views.signQuickpay import signQuickpay
+from members.views.QuickpayCallback import QuickpayCallback
+from members.views.waitinglistView import waitinglistView
+from members.views.paymentGatewayErrorView import paymentGatewayErrorView
+from members.views.departmentView import departmentView
